# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/matieres/{id}": {
            "get": {
                "tags": [
                    "Matieres"
                ],
                "operationId": "App\\Controller\\MatiereController::read",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Matiere"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/matieres/{idModule}": {
            "post": {
                "tags": [
                    "Matieres"
                ],
                "operationId": "App\\Controller\\MatiereController::add",
                "parameters": [
                    {
                        "name": "idModule",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "request": "matieres",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Matiere"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Matiere ajoutée !"
                    },
                    "404": {
                        "description": "Non trouvé..."
                    }
                }
            }
        },
        "/semestres/{id}/matieres": {
            "get": {
                "tags": [
                    "Matieres"
                ],
                "operationId": "App\\Controller\\MatiereController::getMatieresParSemestre",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Matiere"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/assistants/{id}/promotions": {
            "get": {
                "tags": [
                    "Promotions"
                ],
                "operationId": "App\\Controller\\PromotionController::promosParAssistant",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Promotion"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/promotions/{id}/sessions/week/{dateString}": {
            "get": {
                "tags": [
                    "Sessions"
                ],
                "operationId": "App\\Controller\\SessionController::listWeekPourPromo",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id Promotion",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "dateString",
                        "in": "path",
                        "description": "format AAAAMMJJ",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Session"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/sessions/{id}": {
            "get": {
                "tags": [
                    "Sessions"
                ],
                "operationId": "App\\Controller\\SessionController::read",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Session"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/sessions/{idMatiere}": {
            "post": {
                "tags": [
                    "Sessions"
                ],
                "operationId": "App\\Controller\\SessionController::add",
                "parameters": [
                    {
                        "name": "idMatiere",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "request": "sessions",
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Session"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Session ajoutée !"
                    },
                    "404": {
                        "description": "Non trouvée..."
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Matiere": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true"
                    },
                    "nom": {
                        "type": "string"
                    },
                    "coefficient": {
                        "type": "integer"
                    },
                    "sessions": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "$ref": "#/components/schemas/Session/properties/id"
                                },
                                "duree": {
                                    "$ref": "#/components/schemas/Session/properties/duree"
                                }
                            },
                            "type": "object"
                        },
                        "readOnly": "true"
                    },
                    "module": {
                        "properties": {
                            "id": {
                                "$ref": "#/components/schemas/Module/properties/id"
                            },
                            "nom": {
                                "$ref": "#/components/schemas/Module/properties/nom"
                            }
                        },
                        "type": "object",
                        "readOnly": "true"
                    },
                    "notes": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "$ref": "#/components/schemas/Note/properties/id"
                                },
                                "note": {
                                    "$ref": "#/components/schemas/Note/properties/note"
                                },
                                "idEtudiant": {
                                    "$ref": "#/components/schemas/Etudiant/properties/id"
                                }
                            },
                            "type": "object"
                        },
                        "readOnly": "true"
                    },
                    "nombreHeuresAPlacer": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "Promotion": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true"
                    },
                    "nom": {
                        "type": "string"
                    },
                    "semestres": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "$ref": "#/components/schemas/Semestre/properties/id"
                                },
                                "nom": {
                                    "$ref": "#/components/schemas/Semestre/properties/nom"
                                }
                            },
                            "type": "object"
                        },
                        "readOnly": "true"
                    },
                    "formation": {
                        "properties": {
                            "id": {
                                "$ref": "#/components/schemas/Formation/properties/id"
                            },
                            "nom": {
                                "$ref": "#/components/schemas/Formation/properties/nom"
                            }
                        },
                        "type": "object",
                        "readOnly": "true"
                    },
                    "assistant": {
                        "properties": {
                            "id": {
                                "$ref": "#/components/schemas/Assistant/properties/id"
                            },
                            "nom": {
                                "$ref": "#/components/schemas/Personne/properties/nom"
                            },
                            "prenom": {
                                "$ref": "#/components/schemas/Personne/properties/prenom"
                            }
                        },
                        "type": "object",
                        "readOnly": "true"
                    },
                    "nomPromotion": {
                        "type": "string",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            },
...
            "Session": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": "true"
                    },
                    "matiere": {
                        "properties": {
                            "id": {
                                "$ref": "#/components/schemas/Matiere/properties/id"
                            },
                            "nom": {
                                "$ref": "#/components/schemas/Matiere/properties/nom"
                            }
                        },
                        "type": "object",
                        "readOnly": "true"
                    },
                    "type": {
                        "type": "string"
                    },
                    "obligatoire": {
                        "type": "boolean"
                    },
                    "dateDebut": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "dateFin": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "duree": {
                        "type": "string",
                        "format": "date-time",
                        "readOnly": "true"
                    }
                },
                "type": "object"
            }
        }
    }
}